Render 2 electric bogaloo
Project Babylon was a space gun project commissioned by then Iraqi president Saddam Hussein. It involved building a series of "superguns". The design was based on research from the 1960s Project HARP led by the Canadian artillery expert Gerald Bull. There were most likely four different devices in the program.

The project began in 1988; it was halted in 1990 after Bull was assassinated, and parts of the superguns were seized in transit around Europe. The components that remained in Iraq were destroyed by the United Nations after the 1991 Persian Gulf War.