cool front page
Operation Barbarossa[g] was the invasion of the Soviet Union by Nazi Germany and several of its European Axis allies starting on 22 June 1941, during World War II. More than 3.8 million Axis troops invaded the western Soviet Union along a 2,900-kilometer (1,800 mi) front, with the main goal of capturing territory up to a line between Arkhangelsk and Astrakhan (A-A line). The attack became the largest and costliest military offensive in history, with around 10 million combatants taking part[26] and over 8 million casualties by the end of the operation on 5 December 1941.[27][28] It marked a major escalation of World War II, opening the Eastern Front—the largest and deadliest land theatre of war in history—and bringing the Soviet Union into the Allied powers.